refactor: rename GenericLoader to GlobalLoader and refactor useGenericLoader hook
